feat : Create items context for consisting items count
